♻️ refactor: 프로필 이미지 변경 기능 분리

+++ src/main/java/com/grepp/spring/app/controller/api/auth/payload/request/RegisterRequest.java
@@ -11,7 +11,7 @@ public class RegisterRequest {
     private String role;
     private String email;
     private String name;
-    private Long profileImageNumber;
+    private Integer profileImageNumber;
     private String tel;
 
 }
